+++ src/main/java/com/dgu/review/global/configuration/AwsConfig.java
@@ -5,6 +5,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 
 @Configuration
@@ -21,6 +22,13 @@ public class AwsConfig {
     public S3Presigner s3Presigner(Region region) {
         return S3Presigner.builder()
                 .region(region)
+                .credentialsProvider(DefaultCredentialsProvider.create())
+                .build();
+    }
+    @Bean
+    public S3Client s3Client() {
+        return S3Client.builder()
+                .region(Region.of(region))
                 .credentialsProvider(DefaultCredentialsProvider.create())
                 .build();
     }
